requirements.txt
absl-py==1.0.0
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
black==21.12b0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
executing==0.8.2
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gTTS==2.2.3
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.1
ipykernel==6.7.0
ipython==8.0.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jupyter-client==7.1.1
jupyter-core==4.9.1
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
nltk==3.6.7
numpy==1.22.1
oauthlib==3.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prompt-toolkit==3.0.24
protobuf==3.19.3
pure-eval==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pydub==0.25.1
Pygments==2.11.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
stack-data==0.1.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tomli==1.2.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.13.3
zipp==3.7.0
